// solauto-sdk/src/generated/types/automationSettings.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import {
  Serializer,
  array,
  bytes,
  struct,
  u16,
  u64,
  u8,
} from '@metaplex-foundation/umi/serializers';

export type AutomationSettings = {
  targetPeriods: number;
  periodsPassed: number;
  padding1: Array<number>;
  unixStartDate: bigint;
  intervalSeconds: bigint;
  padding: Uint8Array;
};

export type AutomationSettingsArgs = {
  targetPeriods: number;
  periodsPassed: number;
  padding1: Array<number>;
  unixStartDate: number | bigint;
  intervalSeconds: number | bigint;
  padding: Uint8Array;
};

export function getAutomationSettingsSerializer(): Serializer<
  AutomationSettingsArgs,
  AutomationSettings
> {
  return struct<AutomationSettings>(
    [
      ['targetPeriods', u16()],
      ['periodsPassed', u16()],
      ['padding1', array(u8(), { size: 4 })],
      ['unixStartDate', u64()],
      ['intervalSeconds', u64()],
      ['padding', bytes({ size: 32 })],
    ],
    { description: 'AutomationSettings' }
  ) as Serializer<AutomationSettingsArgs, AutomationSettings>;
}